requirements.txt
black>=22.3.0
grpcio
grpcio-tools
grpclib
isort>=5.10.1
jinja2>=3.0.3
protobuf>=4.21.0
psycopg2
pytest>=7.2.1
python-dateutil
pyyaml>=5.1
requests-toolbelt
voluptuous>=0.12.2
yandex-taxi-testsuite>=0.1.18
yandex-taxi-testsuite[postgresql-binary]
yandex-taxi-testsuite[postgresql]
